# api-docs/swagger.yaml
swagger: "2.0"
info:
  title: Book API
  description: A simple API to manage books.
  version: 1.0.0
host: localhost:3000
basePath: /
schemes:
  - http
paths:
  /books:
    get:
      summary: Get all books
      operationId: getBooks
      responses:
        200:
          description: A list of books
          schema:
            type: array
            items:
              type: object
              properties:
                title:
                  type: string
                author:
                  type: string
                published_year:
                  type: integer
                isbn:
                  type: string
                genre:
                  type: string
        404:
          description: No books found
          schema:
            type: object
            properties:
              message:
                type: string
    post:
      summary: Add a new book
      operationId: addBook
      parameters:
        - in: body
          name: book
          required: true
          schema:
            type: object
            properties:
              title:
                type: string
              author:
                type: string
              published_year:
                type: integer
              isbn:
                type: string
              genre:
                type: string
            required:
              - title
              - author
              - published_year
              - isbn
      responses:
        201:
          description: Book added successfully
          schema:
            type: object
            properties:
              message:
                type: string
        400:
          description: Missing required book information
          schema:
            type: object
            properties:
              message:
                type: string
        409:
          description: Book with this ISBN already exists
          schema:
            type: object
            properties:
              message:
                type: string
  /books/search:
    get:
      summary: Search books by author, year, or genre
      operationId: searchBooks
      parameters:
        - name: author
          in: query
          description: Filter books by author
          required: false
          type: string
        - name: year
          in: query
          description: Filter books by publication year
          required: false
          type: integer
        - name: genre
          in: query
          description: Filter books by genre
          required: false
          type: string
      responses:
        200:
          description: A list of filtered books
          schema:
            type: array
            items:
              type: object
              properties:
                title:
                  type: string
                author:
                  type: string
                published_year:
                  type: integer
                isbn:
                  type: string
                genre:
                  type: string
        400:
          description: Invalid year format
          schema:
            type: object
            properties:
              message:
                type: string
        404:
          description: No books found matching the search criteria
          schema:
            type: object
            properties:
              message:
                type: string
  /books/{isbn}:
    delete:
      summary: Delete a book by ISBN
      operationId: deleteBook
      parameters:
        - name: isbn
          in: path
          required: true
          description: The ISBN of the book to delete
          type: string
      responses:
        200:
          description: Book deleted successfully
          schema:
            type: object
            properties:
              message:
                type: string
        404:
          description: Book not found
          schema:
            type: object
            properties:
              message:
                type: string
    put:
      summary: Update a book by ISBN
      operationId: updateBook
      parameters:
        - name: isbn
          in: path
          required: true
          description: The ISBN of the book to update
          type: string
      parameters:
        - in: body
          name: book
          required: true
          schema:
            type: object
            properties:
              title:
                type: string
              author:
                type: string
              published_year:
                type: integer
              genre:
                type: string
      responses:
        200:
          description: Book updated successfully
          schema:
            type: object
            properties:
              message:
                type: string
        404:
          description: Book not found
          schema:
            type: object
            properties:
              message:
                type: string
  /404:
    get:
      summary: Handle 404 errors
      operationId: handle404
      responses:
        404:
          description: Resource not found
          schema:
            type: object
            properties:
              error:
                type: string
                example: "Resource not found"